remove hls config from srs_origin
srs does not really support adaptive bitrate hls chunking on its own. we
really do want to have this feature though, so the transcoding to
different bitrates and chunking to hls/dash/whatever needs to be done
external. ffmpeg could be a very good solution for this, which whould
also allow us to decouple this from the srs nodes and maybe provide
different transcoding config sets via portier/concierge.